requirements.txt
numpy
scipy
torchaudio
soundfile